Nuevos cambios

+++ src/main/java/org/example/proyectocloud/DTO/Admin/Flavors/FlavorRequest.java
@@ -1,12 +1,13 @@
 package org.example.proyectocloud.DTO.Admin.Flavors;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class FlavorRequest implements Serializable {
     private String name;
     private Integer ram;
     private Integer vcpu;
-    private Integer disk;
+    private BigDecimal disk;
     private String type;  // Debe ser "public" o "private"
     private Integer userId;  // Requerido solo para flavors de tipo "private"
 
@@ -35,11 +36,11 @@ public class FlavorRequest implements Serializable {
         this.vcpu = vcpu;
     }
 
-    public Integer getDisk() {
+    public BigDecimal getDisk() {
         return disk;
     }
 
-    public void setDisk(Integer disk) {
+    public void setDisk(BigDecimal disk) {
         this.disk = disk;
     }
